requirements.txt
-i https://pypi.org/simple
appnope==0.1.0 ; platform_system == 'Darwin'
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
click==7.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
flask==1.1.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython-genutils==0.2.0
ipython==7.9.0 ; python_version >= '3.3'
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jinja2==2.10.3
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter==1.0.0
markupsafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0 ; os_name != 'nt'
pycrypto==2.6.1
pygments==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.1
python-decouple==3.1
pyzmq==18.1.0
qtconsole==4.5.5
send2trash==1.5.0
six==1.13.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
